[#41039] linux-client: app: Reinitialize the database when not loaded
When building a clean system image with meta-rdfm, there is no database initialized
at the time of the first boot. As such, the artifact_name and provides are blank.
One way to initialize the database is to install an update for the first time, but
in the state described above this becomes impossible - the installer refuses to
continue when a blank artifact_name is present.

This changes makes RDFM check the database store status when starting - if the DB
is not yet initialized, a dummy artifact_name is set to allow for installation of
proper update images.