[#41944] rdfm-artifact: delta: Simplify ArtifactDelta
Simplify ArtifactDelta by immediately saving the deltas to a file instead
of carrying on the PipeReader chain through even more levels of abstraction.
The `Delta` method now returns a filename to the file containing the raw binary
deltas, it is up to the caller to remove them when no longer required.