[#47449] linux-client: Client daemonization mode
This implements a client daemonization mode, which maintains a connection
to the RDFM management server to provide management and OTA functionality.
The client periodically checks for updates to the device and if one is found,
automatically installs it on the system.
The daemon also allows reverse shell access to the device via the server, and
file upload/download to the device.

Also bumped Dockerfile Debian version to bookworm, as new dependencies require
a more recent version of Go (higher than Go 1.16)