[#47948] server: Implement group management and package assignment
This implements group management functionality (i.e listing, creating, deleting)
to the server API. Each device can be assigned a single group, and a package can
be assigned to a group.

When a group is assigned to any device, deleting it is prohibited. Likewise,
if a package is assigned to any group, attempting to delete it now returns a
409 Conflict status code.