[#47948] server: Group-aware update package checking
Previously, we used simple most-recent-first logic for determining the
next update package. This is now modified to take into consideration
the group a device is assigned to, and the packages assigned to that
specific group.

As a result, the device metadata required for update checking now
also requires passing the device's MAC address.

The update check endpoint was also moved into a separate file, similar
to the other routes.