[#48414] server: Expose package driver in all package endpoints
There was a mismatch between the schema of packages as returned by
the "fetch all packages" endpoint and the "fetch by ID" endpoint.
The former was missing the `driver` field, while they both should
be identical.

Corrected this by adding the `driver` field to the "fetch all packages"
endpoint, and updated the sample responses in the API documentation
(which were also missing the field).