[#48603] server: Update resolution based on package dependency graph
This streamlines the update logic into a separate package resolver.
Instead of using a greedy "find first compatible package" approach, the update
logic now takes into consideration a "target" version that is specified per
the group's policy. If required, intermediate packages will be installed to
ensure that all devices in a group are running the specified version.

This introduces a new concept to the existing package metadata: `requires:`
clauses, which indicate that a package can only be installed if the existing
one contains the specified key/value metadata.