[#49454] server, rdfm-mgmt: Use standard CLI flag convention
The usual naming convention is for long arguments to be prefixed with two dashes, while
shorthands are prefixed with a single dash. This changes the server and the manager to
use this convention. As a result, this commit also updates the old CLI flags in tests
and the documentation to use the new convention.