[#49444] server: Protect all management APIs with enforced authorization
This adds a decorator to mark an endpoint as read-only/read-write. This
is used to provide configurable access levels to all API users. Some users
may have privilege to only read data (example: listing groups, packages, etc.)
while others may also be allowed to modify it (example: creating groups,
packages, etc.).
The decorator also adds scope information to the docstring to be used during
the documentation build.

Incoming requests are expected to contain their authorization token retrieved
from an external authorization server. The RDFM server only validates the
token against it and extracts the scopes granted to the requester to validate
if the user can access the specific endpoint. The authorization server MUST
support the OAuth2 Token Introspection extension to be compatible with RDFM.