[#49444] server: Management API Auth tests
This adds tests for validating the authentication of the management API.
A mock of an authorization server is used to provide the Token Introspection
functionality required by RDFM. These tests verify if scopes provide only
the access levels you'd expect from them (i.e read-only tokens don't allow
you to upload packages). Additionally, added a test that verifies the
behavior of the server when connection to the auth server is impossible
(all requests should fail with 401 Unauthorized).

As adding a decorator is crucial to proper authentication, additionally added
a test that iterates over all registered routes and validates if all of them
have a proper auth decorator added.