[#50470] android-client: Store RDFM key store in internal app storage
This is a slight refactor of the existing device key management
utilities. The RDFM key store is now stored in the internal app
storage available to the RDFM app to work around certain issues
with the default key store under specific system configurations.
The key store is only accessible to the system user and is
protected using a hardcoded password. This password only serves
the purpose of obfuscation, and in the future will be defined
externally as part of the app build (instead of being a constant).