[#50588] android-client: Simplify build-time configuration
This makes the build-time configuration not require copying an actual
source file (Preferences screen XML), but is now provided in a separate
file containing only the required configuration.
The new configuration additionally allows only overlaying certain values
and leaving the rest as defaults, instead of having to provide all of them.