[#76509] linux-client: Add file-backed queue implementation
This implements a persistent queue to be used for storing action
execution requests, as well as their responses. The queue stores simple
binary blobs with no specific structure. Blocking and non-blocking
variants of Dequeue/Enqueue operations are supported.

This also introduces the top-level `actions` package which will store
all actions-related sources. This is done to:
1) Unify how we put modules; telemetry-related sources are already in a
   similar top-level `telemetry` package.
2) Reduce package nesting; major components should be immediately
   visible and not hidden behind multiple layers of packages.
3) Reduce coupling; to help testing components in isolation and not
   within the scope of a single parent package (`daemon`).

Co-authored-by: Mikolaj Klikowicz <mklikowicz@antmicro.com>
[implemented the initial version of the file store]
Signed-off-by: Łukasz Kędziora <lkedziora@antmicro.com>