[#72940] linux-client: Dynamically generate device WS capabilities
Capabilities used in the management WebSocket are meant to indicate the
command set supported by a specific client. They were previously
hardcoded in a configuration file, which meant that different
capabilities could be reported than what the actual daemon supports.
This should be generated on the fly, based on what commands the daemon
can, and is willing to, handle.

Remove the legacy capabilities.toml file and replace it with dynamic
generation of capabilities that the client should support. While the
commands themselves are still handled in handleRequest, this can be
further improved in the future by moving away from the switch-case for
message handling (at which point, command handlers can be registered
dynamically based on configuration).

Signed-off-by: Łukasz Kędziora <lkedziora@antmicro.com>