[#79218] linux-client: Allow specifying regexp to match MAC address interfaces
To make the MAC address interface selection deterministic and not rely
on random ordering, the MacAddressInterfaceRegex config key can be used
to specify a regular expression to use for picking the interface to
source the MAC address identifier from.

Note that this change may introduce deauthorizations of previously
authorized devices if the MacAddressInterfaceRegex is not adjusted
accordingly to match the interface previously used. The logic completely
ignores the IP address now and instead relies on the network interface
name alone, which is more deterministic than relying on the ordering of
network interfaces (which is not guaranteed) and their assigned IP
addresses (which it may not even have).

Signed-off-by: Łukasz Kędziora <lkedziora@antmicro.com>