[#79805] linux-client: Persist autodiscovered MAC address identifier
This adds persistence of the MAC address used as identifier for
communication with the server. Previously, each startup of the client
would re-enumerate the network interface list to match against the
chosen interface to use as source of unique ID. The identifier is now
enumerated only once, on first daemon start.

As a benefit, this additionally enables the ability to provide custom
identifiers for a device, for example custom-formatted serial numbers.

Signed-off-by: Łukasz Kędziora <lkedziora@antmicro.com>