[#87864] devices: linux-client: Allow for a graceful exit before authentication
The daemon used to hang upon a ^C when waiting to be authenticated.
It was due to the fact that device.authenticateDeviceWithServer had been
previously unaware of cancelCtx's existence and it blocked with a
time.Sleep. Modfying it with logic to error out if cancelFunc is called
before a successful authentication allows for a graceful exit.

It is worth to note that this function is not concurrency safe and
should be indirectly called by device.getDeviceToken with a cancelCtx
passed.

This is hacky and it would probably be better if the goroutines/runners
were operating under the assumption that they're already authenticated
with the server.